splits Bigtable.Typed module up into Bigtable.Typed.Reads and Bigtable.Typed.Mutations
